Source-Code/Backend/mqtt-config/application/requirements.txt
bcrypt
mongoengine
paho-mqtt
protobuf
urllib3<2.0
uuid